Fixed proxy credentials not requested for http package source.
When a request was made to a http package source an exception was not
thrown and instead the response is returned with the status code
ProxyAuthenticationRequired. Previously only a request made to a https
package source would request proxy credentials since in this case the
request causes a HttpRequestException to be thrown.