Fix proxy authentication failing on Mono.
When a system wide proxy is configured for all https requests then a
requesting a url such as https://www.nuget.org/api/v2 would fail to
request proxy credentials on Mono so the user would not be prompted.
Instead the request would be retried a few times and then fail.

This only affects https requests. A request made to a http url such
as http://www.nuget.org/api/v2 would request the proxy credentials
from the credential provider.

With Microsoft's .NET Framework a HttpRequestException is thrown with
an InnerException which is an WebException with:

    WebException.Status = ProtocolError
    WebException.Response.StatusCode = ProxyAuthenticationRequired

On Mono a WebException is thrown with:

    WebException.Status = SecureChannelFailure
    WebException.Response = null

To handle this there is now a special case which checks the exception
message for the string 'The remote server returned a 407 status code.'
when running on Mono.

Mono bug report:

https://bugzilla.xamarin.com/show_bug.cgi?id=19594