Workaround Mono problem extracting certain NuGet packages.
When the local timezone is a positive offset from UTC then certain
NuGet packages will fail to extract on Mono with the error:

System.ArgumentOutOfRangeException: The UTC time represented when the offset is applied must be between year 0 and 10,000.
Parameter name: offset
  at System.DateTimeOffset.ValidateDate (DateTime dateTime, TimeSpan offset)
  at System.DateTimeOffset..ctor (DateTime dateTime)
  at System.DateTimeOffset.op_Implicit (DateTime dateTime)
  at System.IO.Compression.ZipArchiveEntry.get_LastWriteTime ()
  at NuGet.Packaging.PackageArchiveReader.CopyFiles (System.String destination, IEnumerable`1 packageFiles, NuGet.Packaging.Core.ExtractPackageFileDelegate extractFile, CancellationToken token)
  at NuGet.Packaging.PackageExtractor.ExtractPackage (NuGet.Packaging.PackageReaderBase packageReader, System.IO.Stream packageStream, NuGet.Packaging.PackagePathResolver packagePathResolver, NuGet.Packaging.PackageExtractionContext packageExtractionContext, CancellationToken token)

This only happens for certain NuGet packages, such as XamarinComponent
1.1.0.32. As a workaround the ArgumentOutOfRangeException is now
handled when reading the LastWriteTime for the zip entry instead of
just when trying to write the file's last write time.